*** We are an amazing ensemble!
This is broken, but it's something we couldn't have done on our own. Each line in pandemic_recovery_batch_test is an idea we had rapidly iterating on what the interface should be for an elegant way to specify column nullability. We tried everyone's ideas, pointed out drawbacks and settled on something (line 46) that none of us could have done on our own.